22.03.24

+++ src/main/resources/shaders/gameubo/fragment.frag
@@ -7,7 +7,7 @@ struct PointLight
     float plG;
     float plB;
     float brightness;
-    int shadowMapId;
+    float shadowMapId;
 };
 
 layout (std140, binding = 0) uniform SunLight {
