+++ src/main/resources/shaders/liquid/fragment.frag
@@ -70,7 +70,7 @@ struct PointLight
     float plG;
     float plB;
     float brightness;
-    int shadowMapId;
+    float shadowMapId;
 };
 
 layout (std140, binding = 0) uniform SunLight {
@@ -222,7 +222,7 @@ vec4 calc_light() {
         float at_base = 1.8 / (bright * 0.5);
         float linear = 2.25 / (bright * 2.75);
         float expo = 0.6 / (bright * 0.25f);
-        vec4 shadow = p.shadowMapId >= 0 ? vec4(calculate_point_light_shadow(point_light_cubemap[p.shadowMapId], out_world_position.xyz, vec3(p.plPosX, p.plPosY, p.plPosZ))) : vec4(1.0);
+        vec4 shadow = p.shadowMapId >= 0 ? vec4(calculate_point_light_shadow(point_light_cubemap[int(p.shadowMapId)], out_world_position.xyz, vec3(p.plPosX, p.plPosY, p.plPosZ))) : vec4(1.0);
         point_light_factor += calc_point_light(p, mv_vertex_pos, normal, at_base, linear, expo, bright) * shadow;
     }
